fixed  Course.CourseID references

+++ ContosoUniversity.Data/Entities/Enrollment.cs
@@ -5,8 +5,8 @@ namespace ContosoUniversity.Data.Entities
 {
     public class Enrollment : BaseEntity
     {
-        public long CourseID { get; set; }
-        public long StudentID { get; set; }
+        public int CourseID { get; set; }
+        public int StudentID { get; set; }
 
         [DisplayFormat(NullDisplayText = "No Grade")]
         public Grade? Grade { get; set; }
